+++ packages/at_app_bundler/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.0
+
+- **Chore**: Updated to support Dart 3
+
 ## 0.2.0
 
   - **Chore**: Updated bundle template build.